LayoutTests/svg/custom/use-nested-missing-target-removed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" onload="runTest()">
    <script type="text/javascript">
    <![CDATA[
    function runTest() {
        if (window.testRunner)
            testRunner.dumpAsText();
        document.getElementById("myText1").id = "myText1";
    }
    ]]>
    </script>
    <text y="30" id="myText" fill="green">PASS</text>
    <use y="30" id="use1" xlink:href="#myText"/>
    <use y="30" xlink:href="#use1"/>
</svg>
