LayoutTests/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE MESSAGE: line 7: TypeError: 'null' is not an object (evaluating 'document.getElementById("myText1").id = "myText1"')
PASS
